record sid clean-up
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@719031 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFSheet.java
@@ -1004,7 +1004,7 @@ public class HSSFSheet implements org.apache.poi.ss.usermodel.Sheet
      * @return hashed password
      */
     public short getPassword() {
-        return getSheet().getPassword().getPassword();
+        return (short)getSheet().getPassword().getPassword();
     }
 
     /**
